New Logs, logs settings, fixes
-User has now the ability to change what  to see on logs channel
-Updated the db schema to allow this new feature
-Updated selectMenu method to allow the select of multiple choices

+++ src/events/messages/messageCreate.js
@@ -28,7 +28,9 @@ module.exports = class messageCreateTracker extends Event {
       const logsChannel = this.client.channels.cache.get(
         fetchGuild.logs.channel
       );
-      if (logsChannel) {
+      const enabledLogs = fetchGuild.logs.enabled;
+
+      if (logsChannel && enabledLogs.includes("moderation")) {
         message.delete();
 
         return logsChannel
